The usual. FIxed some stuff. Probably broke some stuff. Basic drawing woks but wolfie's brushes and broken again.
