+++ static/canvas/canvas.css
@@ -1,6 +1,7 @@
 body{
 padding: 0;
 margin: 0;
+background: url("transparent.png");
 }
 
 #sketch{
@@ -10,7 +11,7 @@ width: 100%;
 }
 
 canvas{
-height: 100%;
 width: 100%;
 position: absolute;
+background: none;
 }
