+++ static/canvas/canvas.html
@@ -14,6 +14,7 @@
     <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
 		<script src="../brushes/aoCanvas.js"></script>
 		<script src="../brushes/brushes.js"></script>
+		<script src="../layers/layers.js"></script>
 		<script src="canvas.js"></script>
   </body>
 </html>
