+++ static/crossWindow.js
@@ -5,5 +5,10 @@ function sendToSibling(iframe, action, data)
 
 function sendToIframe(iframe, action, data)
 {
-  document.getElementById(iframe).contentWindow.postMessage(action + "," + data, '*');
+  var stringData = "";
+  if(data != null)
+    for(var i = 0; i < data.length; i++)
+      stringData+= data[i] + ",";
+
+  document.getElementById(iframe).contentWindow.postMessage(action + "," + stringData, '*');
 }
