+++ static/layers/layers.html
@@ -24,7 +24,7 @@
 
     <!-- JS Scripts -->
     <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
-    <script src="crossWindow.js"></script>
+    <script src="../crossWindow.js"></script>
     <script src="layersUI.js"></script>
   </body>
 </html>
